refactor(app): add launcher phase entrypoint poc
