+++ app/launcher/precompile_app.jl
@@ -1,0 +1,4 @@
+using JRTLauncherApp
+
+JRTLauncherApp.real_main(["help"])
+JRTLauncherApp.real_main(["phase", "--help"])
